Fix readme local demo (#85)
* add "optional: true" to packages in package-lock.json

Running npm install on the project will result in modification of
package-lock.json. I ran npm install and the file was modified.

* Add path to local web server demo to README.md

* Fix port number and add --open alternative

Removed the port number since it can be different. Added another line to
show you can alternatively run the server with --open and add /demo to
the URL.