Refresh on current topic (#91)
* When the HREF changes, it should set the new url without reloading, allows for back and forth but it will not work when the app is served in a IFrame

* Fix refresh on current topic

The _initialize(href) function will now set the URL of the page to
reflect the new activity being viewed each time a new activity is
loaded. With this, when the page is refreshed, ASV will load the
last activity that was viewed.

* Removing comments

These comments were written by lawrence to explain the work in progress
and are no longer needed.

* Resolve lint issues

There were linting problems. These changes resolve those problems.

* Remove line that sets value of referrer

After looking at other components of the code, it turned out that we're
not using the value of HTTPReferrer for anything which is why this line
is being removed.